==> R/2. SfC_generating_ipw.Rmd <==
---
title: "Data Cleaning for Skills for Care Paper"
author: "Meng Ye"
date: "7/7/2024"
output:
  word_document: default
  pdf_document: default
  html_document: default
---

## Read me

This is the data cleaning file for generating ipw for the CQC locations selecting into Skills for Care data set. The overall steps is as follows

1. Take the anonymized combined CQC-SfC data, generate a dummy variable based on whether SfC establishment_id is NA or not, `select_in` 
2. Filter out the Overall rating rows and calculate the ipw at the location level
3. Run the logit regression model of `select_in` and calculate the ipw
4. Derive the sub-data frame of and `establishment_id` and `ipw`
5. Import the cleaned SfC prepared in Step 1
6. Merge the ipw data-frame to the cleaned SfC data


## Setting up

```{r load package, message=FALSE, warning=FALSE}
library(tidyverse) # package for data cleaning and plotting
library(readxl) # package for reading excel file
library(broom) # extracting model summary as data frame
library(modelsummary) # deriving model tables
library(scales) # label percent
library(lubridate) # working with dates
library(here) # working with diectory
library(gt) # grammar of tables
```

```{r contr}
# environment setup to run ordered logit properly
options(contrasts = rep("contr.treatment", 2))
```

## Import data 

The orignial joint data returned by Skills for care with anonymous providers

```{r import data, message=FALSE, warning=FALSE}
#import the raw data file
cqc_skills <- read_excel(here("data","SfC_CQC.xlsx"))
```

# data set of rows of CQC with cleaned Skills for Care data 

```{r import cleaned cqc data, message=FALSE, warning=FALSE}
#import the cleaned skills for care data
cleaned_cqc_skills <- read_rds(here("data","cqc_sfc_cleaned.rds"))
```


```{r}
# check service group (by Skills for Care) distribution 
cleaned_cqc_skills |>
  group_by(service_type_selected) |>
  summarize(count = n())
```


## Data Cleaning for CQC_skills combined data set 

```{r}
# prepare the full CQC data for the logit regression
cqc <- cqc_skills %>% 
  mutate(select_in = ifelse(!is.na(establishmentid), 1, 0)) %>% 
    mutate(inherited = ifelse(inherited == "Y", TRUE, FALSE),
         rating = recode(rating, 
                         "Insufficient evidence to rate" = "NA",
                         "Requires improvement" = "Req improv")) %>% 
  # set the order of the values in the factors 
  mutate(form = fct_relevel(form, "FPO"),
         
  # assume the order of the ratings as follows but need to double check with the source 
         rating = ordered(rating, levels = c("Inadequate","Req improv", "Good", "Outstanding"))) %>%
  
  # adding the rating data coded as numerical
  mutate(rating_num = case_when(rating == "Inadequate" ~ 1,
                                rating == "Req improv" ~ 2,
                                rating == "Good" ~ 3,
                                rating == "Outstanding" ~ 4)) %>% 
  
  mutate(category = case_when(primary_cat == "Community based adult social care services" ~ "community",
                              primary_cat == "Residential social care" ~ "residential",
                              TRUE ~ as.character(NA)),

         # deriving year column and dummy variable for before_covid
         year = year(publication_date),
         Year = factor(year),
         during_covid = ifelse(year >= 2020, TRUE, FALSE),
         before_covid = ifelse(year <= 2019, TRUE, FALSE))

head(cqc)
```


## derive overall rating sub data frame

```{r}
cqc_overall <- cqc %>%
  filter(domain == "Overall")
```

## Run logit model and calculate ipw

```{r}
model_select <- glm(select_in ~ form + category + region + 
                    publication_date + inherited + rating_num,
                    data = cqc_overall,
                    family = binomial(link = "logit"))

tidy(model_select)
```



## Calculate IPW


```{r}
ipw_data <- augment_columns(model_select,
                            cqc_overall,
                            type.predict = "response") %>% 
            rename(propensity = .fitted) %>% 
            mutate(ipw = (select_in / propensity) + (1 - select_in) / (1 - propensity)) %>% 
            select(establishmentid, ipw) %>% 
            filter(establishmentid != "NA") %>% 
            group_by(establishmentid) %>% 
            summarise(ipw = mean(ipw)) %>% 
            ungroup()

nrow(ipw_data)
```


```{r}
augment_columns(model_select,
                            cqc_overall,
                            type.predict = "response") %>% 
            rename(propensity = .fitted) %>% 
            mutate(ipw = (select_in / propensity) + (1 - select_in) / (1 - propensity)) %>% 
            select(establishmentid, ipw) %>% 
            filter(establishmentid != "NA") %>% 
            summarise(count = n(),
                      unic = n_distinct(establishmentid)) 
```

The total count is for rating unit, not for establishment. The Skills for Care data for the same location-id will be duplicated for different measuring units. The unique establishment id is 6505. Only 2 duplicates. 


```{r checking repeated measureing unit}
cqc |>
 group_by(service_group) |>
 summarize(count = n())
```



## Merge the ipw data into the cleaned, CQC-Skills for care data

```{r}
cleaned_sfc_ipw <- cleaned_cqc_skills %>% 
  left_join(ipw_data, by = "establishmentid") 
```


## Save data

```{r save data}
 write_rds(cleaned_sfc_ipw, file = here("data","cleaned_sfc_ipw.rds"))  
```


## Data overview - derive CQC-SFC compare tables

### build a data frame to derive the tables

```{r}
compare_data <- cqc %>% 
            left_join(ipw_data, by = "establishmentid") %>% 
            mutate(SfC = ifelse(select_in == 1, TRUE, FALSE),
                   weighted_rating_num = rating_num * ipw) %>% 
   filter(!is.na(form)& !is.na(rating))
```


```{r message=FALSE, warning=FALSE}
sfc_count <- compare_data %>%
  filter(select_in == TRUE) %>% 
  group_by(form, rating) %>%
  summarize(SfC = n(),
            SfC_weighted = sum(ipw, na.rm = TRUE)) %>% 
  ungroup() %>% 
  filter(form !="NA")

sfc_count
```

```{r}
cqc_count <- compare_data %>%
  group_by(form, rating) %>%
  summarize(CQC = n()) %>% 
  left_join(sfc_count, by = c("form", "rating")) %>% 
  ungroup() %>% 
  filter(form!="NA") %>%
  mutate(ratio = SfC/CQC) 
cqc_count
```

```{r}
sfc_joint <- compare_data %>%
  filter(select_in == TRUE) %>% 
  group_by(form) %>%
  summarize(SfC = n(),
            SfC_weighted = sum(ipw, na.rm = TRUE))
```

```{r}
cqc_subtotal <- compare_data %>%
  group_by(form) %>%
  summarize(CQC = n()) %>% 
  left_join(sfc_joint, by = "form") %>% 
  ungroup() %>% 
  filter(form!="NA") %>%
  mutate(ratio = SfC/CQC) %>% 
  mutate(rating = "Sub-total")
cqc_subtotal
```
### stack two data frames together

```{r}
compare_table <- bind_rows(cqc_count, cqc_subtotal) %>% 
  mutate(rating = factor(rating, levels = c("Outstanding", "Good", "Req improv", "Inadequate", "Sub-total"), ordered = TRUE)) %>% 
  arrange(form, rating) %>% 
  mutate(ratio = scales::percent(ratio, scale = 100, accuracy = 0.1)) %>% 
  mutate(SfC_weighted= scales::label_number(big.mark = ",", accuracy = 1)(SfC_weighted),
         CQC= scales::label_number(big.mark = ",", accuracy = 1)(CQC),
         SfC= scales::label_number(big.mark = ",",accuracy = 1)(SfC)) %>% 
  gt()


  

compare_table
```
