==> R/3. skills-for-care-Draft1.Rmd <==
---
title: "Skills-for-care analysis Draft 1"
author: "Meng Ye"
date: "2024-07-31"
output:
  word_document: default
  html_document: default
---

7/31/2024 Add size (total stuff) control to the models.

Updated with newly derived variables.




```{r contr}
# environment setup to run ordered logit properly
options(contrasts = rep("contr.treatment", 2),
        scipen = 999)
```


This chunk loads all the packages to use

```{r load package, message=FALSE, warning=FALSE}
#packages for ordered logit
library(tidyverse) # package for data cleaning and plotting
library(readxl) # package for reading excel file
library(broom) # extracting model summary as data frame
library(modelsummary) # deriving model tables
library(scales) # label percent
library(lubridate) # working with dates
library(marginaleffects) #to calculate marginal effects
library(ordinal) # package for ordinal logit regression
library(here) # specifying directory 
```



## Data Preparation

First, import the sampled and coded data set

```{r import data, message=FALSE, warning=FALSE}
#import the raw data file
cqc_skills <- read_rds(here("data","cleaned_sfc_ipw.rds")) 

```


```{r}
cqc_skills %>% 
  mutate(overall = ifelse(domain == "Overall", 1, 0)) %>% 
  group_by(form) %>% 
  summarize(count_rating_unit = sum(overall),
            count_rating = n())
```
```{r}
nrow(cqc_skills)
```



```{r}
summary(cqc_skills[, c("turnover", "turnover0", "staff_level", "staff_level0")])
```
where "turnover0" and "staff_level0" denote the older/crude way to calculate the variables. 

```{r}
cqc_skills %>% 
  group_by(form) %>% 
  summarise(count_establishment = n_distinct(establishmentid),
            count_rating = n())
```




```{r}
datasummary(inherited + during_covid + care_home ~ Percent(), data = cqc_skills)
```


## Modelis Used / Reported

```{r }
model_order_weighted <- clm(rating ~ form + care_home + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                link = "logit",
                weights = ipw)

model_staff_weighted <- lm(staff_level ~ form + care_home + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                weights = ipw)

staff_direct_order_weighted <- clm(rating ~ form + staff_level + care_home + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                link = "logit",
                weights = ipw)

model_turnover_weighted <- lm(turnover ~ form + care_home + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                weights = ipw)

turnover_direct_order_weighted <- clm(rating ~ form + turnover + care_home + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                link = "logit",
                weights = ipw)
```



### results table for new staffing variables 

```{r}
new_models <- modelsummary(list("total effect" = model_order_weighted,  
                                "staffing by sector"= model_staff_weighted,
                                "direct effect: staffing control" = staff_direct_order_weighted,
                                "turnover by sector"= model_turnover_weighted, 
                                "direct effect: turnover control" = turnover_direct_order_weighted),
                               coef_omit = "region", exponentiate = F,
                               statistic = "({p.value}) {stars}")
new_models
```

```{r}
new_models_exp <- modelsummary(list("total effect" = model_order_weighted,  
                                "staffing by sector"= model_staff_weighted,
                                "direct effect: staffing control" = staff_direct_order_weighted,
                                "turnover by sector"= model_turnover_weighted, 
                                "direct effect: turnover control" = turnover_direct_order_weighted),
                               coef_omit = "region|\\(Intercept\\)", exponentiate = T,
                               statistic = "({p.value}) {stars}")
new_models_exp
```


## Derive coefficient plots

```{r}

# Tidy the models and filter out the threshold terms
staff_tidied <- tidy(model_staff_weighted, conf.int = TRUE) %>%
  filter(!term %in% c("(Intercept)", "Inadequate|Req improv", "Req improv|Good", "Good|Outstanding")) %>%
  filter(!str_detect(term, "^region"))

turnover_tidied <- tidy(model_turnover_weighted, conf.int = TRUE) %>%
  filter(!term %in% c("(Intercept)", "Inadequate|Req improv", "Req improv|Good", "Good|Outstanding")) %>%
  filter(!str_detect(term, "^region"))

order_tidied <- tidy(model_order_weighted, conf.int = TRUE) %>%
  filter(!term %in% c("(Intercept)", "Inadequate|Req improv", "Req improv|Good", "Good|Outstanding")) %>%
  filter(!str_detect(term, "^region")) %>%
  mutate(model = "Total effects")

staff_direct_tidied <- tidy(staff_direct_order_weighted, conf.int = TRUE) %>%
  filter(!term %in% c("(Intercept)", "Inadequate|Req improv", "Req improv|Good", "Good|Outstanding")) %>%
  filter(!str_detect(term, "^region")) %>%
  mutate(model = "Direct effects")

turnover_direct_tidied <- tidy(turnover_direct_order_weighted, conf.int = TRUE) %>%
  filter(!term %in% c("(Intercept)", "Inadequate|Req improv", "Req improv|Good", "Good|Outstanding")) %>%
  filter(!str_detect(term, "^region")) %>%
  mutate(model = "Direct Effect")

# Combine order and staff direct models
combined_staff_tidied <- bind_rows(order_tidied, staff_direct_tidied)


# Combine total effect and direct effect models
combined_turnover_tidied <- bind_rows(order_tidied, turnover_direct_tidied)

```



```{r}
ggplot(staff_tidied, aes(x = estimate, y = term)) +
  geom_vline(xintercept = 0, color = "red", linetype = "dotted") +
  geom_pointrange(aes(xmin = conf.low, xmax = conf.high)) + 
  labs(x = "Coefficient estimate", y = NULL) +
  theme_minimal() +
  ggtitle("Staffing by Sector Model Coefficient Plot")

```

```{r}
ggplot(turnover_tidied, aes(x = estimate, y = term)) +
  geom_vline(xintercept = 0, color = "red", linetype = "dotted") +
  geom_pointrange(aes(xmin = conf.low, xmax = conf.high)) + 
  labs(x = "Coefficient estimate", y = NULL) +
  theme_minimal() +
  ggtitle("Turnover by Sector Model Coefficient Plot")

```

```{r}
ggplot(combined_staff_tidied, aes(x = estimate, y = term, color = model, shape = model)) +
  geom_vline(xintercept = 0, color = "red", linetype = "dotted") +
  geom_pointrange(aes(xmin = conf.low, xmax = conf.high), position = position_dodge(width = 0.5)) + 
  labs(x = "Coefficient estimate", y = NULL) +
  theme_minimal() +
  ggtitle("Staffing Level: total effects and direct effects")

```





```{r}
ggplot(combined_turnover_tidied, aes(x = estimate, y = term, color = model, shape = model)) +
  geom_vline(xintercept = 0, color = "red", linetype = "dotted") +
  geom_pointrange(aes(xmin = conf.low, xmax = conf.high), position = position_dodge(width = 0.5)) + 
  labs(x = "Coefficient estimate", y = NULL) +
  theme_minimal() +
  ggtitle("Turnover: Total Effects and Direct Effects")

```


## Additional models

### Modeling on rating with old staffing variables and different category

```{r }
model_staff_weighted0 <- lm(staff_level0 ~ form + service_type_selected + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                weights = ipw)

staff_direct_order_weighted0 <- clm(rating ~ form + staff_level0 + service_type_selected + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                link = "logit",
                weights = ipw)

model_turnover_weighted0 <- lm(turnover0 ~ form + service_type_selected + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                weights = ipw)

turnover_direct_order_weighted0 <- clm(rating ~ form + turnover0 + service_type_selected + age + totalstaff + region + during_covid,
                data = filter(cqc_skills, domain == "Overall"),
                link = "logit",
                weights = ipw)
```




### results table for oldstaffing variables


```{r}
old_models <- modelsummary(list("full weighted" = model_order_weighted,  
                                "staffing weighted"= model_staff_weighted0,
                                "staff direct weighted" = staff_direct_order_weighted0,
                                "turnover weighted"= model_turnover_weighted0, 
                                "turnover direct weighted" = turnover_direct_order_weighted0),
                               coef_omit = "region", exponentiate = F,
                               statistic = "({p.value}) {stars}")
old_models
```


## Models for other domains (currently without totalstaff control)



"Overall","Safe","Effective", "Caring", "Well-led", "Responsive"

### "Safe" domain 

```{r }
model_order_weighted_safe <- clm(rating ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Safe"),
                link = "logit",
                weights = ipw)

model_staff_weighted_safe <- lm(staff_level ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Safe"),
                weights = ipw)

staff_direct_order_weighted_safe <- clm(rating ~ form + staff_level + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Safe"),
                link = "logit",
                weights = ipw)

model_turnover_weighted_safe <- lm(turnover ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Safe"),
                weights = ipw)

turnover_direct_order_weighted_safe <- clm(rating ~ form + turnover + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Safe"),
                link = "logit",
                weights = ipw)
```

```{r}
safe_models <- modelsummary(list("full weighted" = model_order_weighted_safe,  
                                "staffing weighted"= model_staff_weighted_safe,
                                "staff direct weighted" = staff_direct_order_weighted_safe,
                                "turnover weighted"= model_turnover_weighted_safe, 
                                "turnover direct weighted" = turnover_direct_order_weighted_safe),
                               coef_omit = "region", exponentiate = F,
                               statistic = "({p.value}) {stars}")
safe_models
```

### "Effective domain"
```{r }
model_order_weighted_effective <- clm(rating ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Effective"),
                link = "logit",
                weights = ipw)

model_staff_weighted_effective <- lm(staff_level ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Effective"),
                weights = ipw)

staff_direct_order_weighted_effective <- clm(rating ~ form + staff_level + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Effective"),
                link = "logit",
                weights = ipw)

model_turnover_weighted_effective <- lm(turnover ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Effective"),
                weights = ipw)

turnover_direct_order_weighted_effective <- clm(rating ~ form + turnover + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Effective"),
                link = "logit",
                weights = ipw)
```


```{r}
effective_models <- modelsummary(list("full weighted" = model_order_weighted_effective,  
                                "staffing weighted"= model_staff_weighted_effective,
                                "staff direct weighted" = staff_direct_order_weighted_effective,
                                "turnover weighted"= model_turnover_weighted_effective, 
                                "turnover direct weighted" = turnover_direct_order_weighted_effective),
                               coef_omit = "region", exponentiate = F,
                               statistic = "({p.value}) {stars}")
effective_models
```


### "Caring" domain 

```{r }
model_order_weighted_caring <- clm(rating ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Caring"),
                link = "logit",
                weights = ipw)

model_staff_weighted_caring <- lm(staff_level ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Caring"),
                weights = ipw)

staff_direct_order_weighted_caring <- clm(rating ~ form + staff_level + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Caring"),
                link = "logit",
                weights = ipw)

model_turnover_weighted_caring <- lm(turnover ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Caring"),
                weights = ipw)

turnover_direct_order_weighted_caring <- clm(rating ~ form + turnover + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Caring"),
                link = "logit",
                weights = ipw)
```


```{r}
caring_models <- modelsummary(list("full weighted" = model_order_weighted_caring,  
                                "staffing weighted"= model_staff_weighted_caring,
                                "staff direct weighted" = staff_direct_order_weighted_caring,
                                "turnover weighted"= model_turnover_weighted_caring, 
                                "turnover direct weighted" = turnover_direct_order_weighted_caring),
                               coef_omit = "region", exponentiate = F,
                               statistic = "({p.value}) {stars}")
caring_models
```

### "Well-led" domain

```{r }
model_order_weighted_well_led <- clm(rating ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Well-led"),
                link = "logit",
                weights = ipw)

model_staff_weighted_well_led <- lm(staff_level ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Well-led"),
                weights = ipw)

staff_direct_order_weighted_well_led <- clm(rating ~ form + staff_level + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Well-led"),
                link = "logit",
                weights = ipw)

model_turnover_weighted_well_led <- lm(turnover ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Well-led"),
                weights = ipw)

turnover_direct_order_weighted_well_led <- clm(rating ~ form + turnover + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Well-led"),
                link = "logit",
                weights = ipw)
```


```{r}
well_led_models <- modelsummary(list("total effect" = model_order_weighted_well_led,  
                                "staffing by sector"= model_staff_weighted_well_led,
                                "direct effect: staffing control" = staff_direct_order_weighted_well_led,
                                "turnover by sector"= model_turnover_weighted_well_led, 
                                "direct effect: turnover control" = turnover_direct_order_weighted_well_led),
                               coef_omit = "region", exponentiate = F,
                               statistic = "({p.value}) {stars}")
well_led_models
```


### "Responsive" domain 

```{r }
model_order_weighted_responsive <- clm(rating ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Responsive"),
                link = "logit",
                weights = ipw)

model_staff_weighted_responsive <- lm(staff_level ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Responsive"),
                weights = ipw)

staff_direct_order_weighted_responsive <- clm(rating ~ form + staff_level + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Responsive"),
                link = "logit",
                weights = ipw)

model_turnover_weighted_responsive <- lm(turnover ~ form + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Responsive"),
                weights = ipw)

turnover_direct_order_weighted_responsive <- clm(rating ~ form + turnover + service_type_selected + age + region + during_covid,
                data = filter(cqc_skills, domain == "Responsive"),
                link = "logit",
                weights = ipw)
```


```{r}
responsive_models <- modelsummary(list("full weighted" = model_order_weighted_responsive,  
                                "staffing weighted"= model_staff_weighted_responsive,
                                "staff direct weighted" = staff_direct_order_weighted_responsive,
                                "turnover weighted"= model_turnover_weighted_responsive, 
                                "turnover direct weighted" = turnover_direct_order_weighted_responsive),
                               coef_omit = "region", exponentiate = F,
                               statistic = "({p.value}) {stars}")
responsive_models
```
